scripts_python/requirements.txt
pygame==2.5.1
pylibftdi==0.21.0
pyserial==3.5